+++ gradle.properties
@@ -6,8 +6,8 @@ maven_group=io.github.llamarama.team
 archives_base_name=Llamarama
 # Minecraft/Fabric Loader deps
 minecraft_version=1.18.2
-yarn_mappings=1.18.2+build.2
+yarn_mappings=1.18.2+build.3
 loader_version=0.13.3
 #Fabric api
-fabric_version=0.48.0+1.18.2
+fabric_version=0.51.1+1.18.2
