+++ src/main/java/io/github/llamarama/team/common/block/blockentity/ModBlockEntityTypes.java
@@ -16,7 +16,9 @@ public final class ModBlockEntityTypes {
 
     @SuppressWarnings("EmptyMethod")
     public static void init() {
-    }    public static final BlockEntityType<LlamaWoolBedBlockEntity> LLAMA_WOOL_BED =
+    }
+
+    public static final BlockEntityType<LlamaWoolBedBlockEntity> LLAMA_WOOL_BED =
         create("llama_wool_bed", LlamaWoolBedBlockEntity::new, ModBlocks.LLAMA_WOOL_BED);
 
     private static <T extends BlockEntity> BlockEntityType<T> create(String id,
@@ -28,8 +30,6 @@ public final class ModBlockEntityTypes {
     private static <T extends BlockEntity> BlockEntityType<T> register(String id, BlockEntityType<T> blockEntityType) {
         return Registry.register(Registry.BLOCK_ENTITY_TYPE, IdBuilder.of(id), blockEntityType);
     }
-
-
 
 
 }
