+++ src/main/java/io/github/llamarama/team/common/entity/ai/goal/SpitRevengeGoal.java
@@ -1,6 +1,6 @@
 package io.github.llamarama.team.common.entity.ai.goal;
 
-import io.github.llamarama.team.mixins.AccessorLlamaEntity;
+import io.github.llamarama.team.mixin.AccessorLlamaEntity;
 import net.minecraft.entity.ai.goal.RevengeGoal;
 import net.minecraft.entity.mob.PathAwareEntity;
 import net.minecraft.entity.passive.LlamaEntity;
