+++ src/main/java/io/github/llamarama/team/common/util/PosUtilities.java
@@ -1,9 +1,7 @@
 package io.github.llamarama.team.common.util;
 
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.Vec3d;
-import net.minecraft.util.math.Vec3i;
 import net.minecraft.world.Heightmap;
 import net.minecraft.world.World;
 import org.jetbrains.annotations.NotNull;
