+++ src/main/java/io/github/llamarama/team/mixin/MixinServerWorld.java
@@ -1,6 +1,7 @@
-package io.github.llamarama.team.mixins;
+package io.github.llamarama.team.mixin;
 
 import io.github.llamarama.team.common.entity.spawn.CaravanTraderSpawnFactory;
+import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.util.profiler.Profiler;
 import net.minecraft.util.registry.RegistryEntry;
@@ -9,6 +10,8 @@ import net.minecraft.world.MutableWorldProperties;
 import net.minecraft.world.StructureWorldAccess;
 import net.minecraft.world.World;
 import net.minecraft.world.dimension.DimensionType;
+import net.minecraft.world.level.ServerWorldProperties;
+import net.minecraft.world.level.storage.LevelStorage;
 import net.minecraft.world.spawner.Spawner;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.asm.mixin.Mixin;
@@ -17,6 +20,7 @@ import org.spongepowered.asm.mixin.injection.ModifyVariable;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.Executor;
 import java.util.function.Supplier;
 
 @Mixin(ServerWorld.class)
@@ -31,8 +35,8 @@ public abstract class MixinServerWorld extends World implements StructureWorldAc
         at = @At("HEAD"),
         argsOnly = true
     )
-    private static @NotNull List<Spawner> modifySpawnsList(List<Spawner> list) {
-        if (!list.isEmpty()) {
+    private static @NotNull List<Spawner> modifySpawnsList(List<Spawner> list, MinecraftServer server, Executor workerExecutor, LevelStorage.Session session, ServerWorldProperties properties, RegistryKey<World> worldKey) {
+        if (worldKey.getValue().equals(World.OVERWORLD.getValue()) && !list.isEmpty()) {
             var newList = new ArrayList<>(list);
             newList.add(new CaravanTraderSpawnFactory());
