+++ src/main/resources/llamarama.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "io.github.llamarama.team.mixins",
+  "package": "io.github.llamarama.team.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
     "AccessorLlamaEntity",
